mise en place du bonnus l'audio

+++ GestionSupermarche/DTO/TempsTravailDetailDto.cs
@@ -6,6 +6,6 @@
         public string NomEmploye { get; set; }
         public string NomRayon { get; set; }
         public DateTime Date { get; set; }
-        public double Temps { get; set; }
+        public int Temps { get; set; }
     }
 }
