+++ GestionSupermarche/DTO/TempsTravailDto.cs
@@ -5,6 +5,6 @@
         public int IdTempsTravail { get; set; }
         public string NomRayon { get; set; }
         public DateTime Date { get; set; }
-        public double Temps { get; set; }
+        public int Temps { get; set; }
     }
 }
